requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
colorama==0.3.9
Django==2.1.1
djangorestframework==3.8.2
more-itertools==4.3.0
pluggy==0.7.1
py==1.6.0
pytest==3.8.0
pytz==2018.5
six==1.11.0
